Renaming of the outdated classifier module. This is destined to be deleted once everything has been fully tested/integrated.
